Add: base of theme. CHONK COMMIT
Nicely integrated theme with my old sites design. It uses hugo with 2
themes. The themes are PaperMod and hugo-notice. PaperMod is the base
theme and does most of the pages. I used to use FixIt but it has some
very opinionated choices that I dont agree with and no way to configure
it nicer. PaperMod has more configurable settings and more sensible
opinions IMO.

hugo-notice is for notices, tips, etc. To make it work with the theme
but not the mode (dark, light), I had to edit the css to always be the
dark mode fonts. hugo-notices all have light backgrounds. I had to hard
code the colours because the FixIt has hardcoded themes instead of
extracted variable font themes that could be configured.

I wrote a few posts and theyre *ok*, I am workin on rewriting them to be
better.

Josh Boutros. He is a great friend and an exceptional graphic designer.
He made the typerwriter icon and he did an exceptional job. I am very
happy to have is as a friend and I am grateful for these favours. He did
a very very good job.

The typewriter icon links to the posts(maybe gonna change to archives).
The other change to go along with this is the about page. If you hover
over my photo, it rises and if you click it, it goes to the about me
page.

I think thats it. This is a big commit.